mavlink: move TRAJECTORY_REPRESENTATION_WAYPOINTS to separate streams header
